feat(exceptions): adiciona tratamento global e ajustes nos controllers e services

+++ src/main/java/com/gft/dicionario/dicionario_api/service/LabelService.java
@@ -2,6 +2,9 @@ package com.gft.dicionario.dicionario_api.service;
 
 
 import com.gft.dicionario.dicionario_api.entities.Label;
+import com.gft.dicionario.dicionario_api.entities.Word;
+import com.gft.dicionario.dicionario_api.exceptions.BadRequestException;
+import com.gft.dicionario.dicionario_api.exceptions.ResourceNotFoundException;
 import com.gft.dicionario.dicionario_api.repositories.LabelRepository;
 import org.springframework.stereotype.Service;
 
@@ -22,19 +25,44 @@ public class LabelService {
         return labelRepository.findAll();
     }
 
-    public Optional<Label> findById(Long id){
-        return labelRepository.findById(id);
+    public Label findById(Long id){
+        return labelRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(
+                "Etiqueta com ID " + id + " não encontrada."
+        ));
     }
 
     public Label save(Label label){
+        if (label == null) {
+            throw new BadRequestException("O corpo da requisição não pode ser nulo.");
+        }
+
+        if (label.getLabelName() == null || label.getLabelName().isBlank()) {
+            throw new BadRequestException("O campo não pode ser vazio.");
+        }
         return labelRepository.save(label);
     }
 
     public void deleteById(Long id){
+        if (!labelRepository.existsById(id)) {
+            throw new ResourceNotFoundException(
+                    "Não foi possível deletar. Etiqueta com ID " + id + " não encontrada.");
+        }
+
         labelRepository.deleteById(id);
     }
 
     public Set<Label> findByWord(String term){
+        if (term == null || term.isBlank()) {
+            throw new BadRequestException("O parâmetro 'labelName' não pode ser vazio.");
+        }
+
+        Set<Label> words = labelRepository.findByWordList_Term(term);
+
+        if (words == null || words.isEmpty()) {
+            throw new ResourceNotFoundException(
+                    "Nenhuma palavra encontrada com o label '" + term + "'.");
+        }
+
         return labelRepository.findByWordList_Term(term);
     }
 }
